third_party/blink/web_tests/external/wpt/css/css-images/image-set/image-set-computed.sub-expected.txt

This is a testharness.js-based test.
FAIL Property background-image value 'image-set(url(http://web-platform.test/example.png) 1x)' assert_equals: expected "image-set(url(\"http://web-platform.test/example.png\") 1dppx)" but got "image-set(url(\"http://web-platform.test/example.png\") 1x)"
FAIL Property background-image value '-webkit-image-set(url(http://web-platform.test/example.png) 1x)' assert_equals: expected "image-set(url(\"http://web-platform.test/example.png\") 1dppx)" but got "image-set(url(\"http://web-platform.test/example.png\") 1x)"
FAIL Property background-image value 'image-set(url(http://web-platform.test/example.png) 1x, 'http://web-platform.test/example.png' 2x)' assert_equals: expected "image-set(url(\"http://web-platform.test/example.png\") 1dppx, url(\"http://web-platform.test/example.png\") 2dppx)" but got "image-set(url(\"http://web-platform.test/example.png\") 1x, url(\"http://web-platform.test/example.png\") 2x)"
FAIL Property background-image value '-webkit-image-set(url(http://web-platform.test/example.png) 1x, 'http://web-platform.test/example.png' 2x)' assert_equals: expected "image-set(url(\"http://web-platform.test/example.png\") 1dppx, url(\"http://web-platform.test/example.png\") 2dppx)" but got "image-set(url(\"http://web-platform.test/example.png\") 1x, url(\"http://web-platform.test/example.png\") 2x)"
FAIL Property background-image value 'image-set(url("http://web-platform.test/example.png") 1dppx)' assert_true: 'image-set(url("http://web-platform.test/example.png") 1dppx)' is a supported value for background-image. expected true got false
FAIL Property background-image value '-webkit-image-set(url("http://web-platform.test/example.png") 1dppx)' assert_true: '-webkit-image-set(url("http://web-platform.test/example.png") 1dppx)' is a supported value for background-image. expected true got false
FAIL Property background-image value 'image-set(url(http://web-platform.test/example.png) 48dpi)' assert_true: 'image-set(url(http://web-platform.test/example.png) 48dpi)' is a supported value for background-image. expected true got false
FAIL Property background-image value '-webkit-image-set(url(http://web-platform.test/example.png) 48dpi)' assert_true: '-webkit-image-set(url(http://web-platform.test/example.png) 48dpi)' is a supported value for background-image. expected true got false
FAIL Property background-image value 'image-set(url(http://web-platform.test/example.png) 2400dpcm, 'http://web-platform.test/example.png' 2x)' assert_true: 'image-set(url(http://web-platform.test/example.png) 2400dpcm, 'http://web-platform.test/example.png' 2x)' is a supported value for background-image. expected true got false
FAIL Property background-image value '-webkit-image-set(url(http://web-platform.test/example.png) 2400dpcm, 'http://web-platform.test/example.png' 2x)' assert_true: '-webkit-image-set(url(http://web-platform.test/example.png) 2400dpcm, 'http://web-platform.test/example.png' 2x)' is a supported value for background-image. expected true got false
FAIL Property background-image value 'image-set('http://web-platform.test/example.jpeg' 240dpi, url(http://web-platform.test/example.png) 3.5x)' assert_true: 'image-set('http://web-platform.test/example.jpeg' 240dpi, url(http://web-platform.test/example.png) 3.5x)' is a supported value for background-image. expected true got false
FAIL Property background-image value '-webkit-image-set('http://web-platform.test/example.jpeg' 240dpi, url(http://web-platform.test/example.png) 3.5x)' assert_true: '-webkit-image-set('http://web-platform.test/example.jpeg' 240dpi, url(http://web-platform.test/example.png) 3.5x)' is a supported value for background-image. expected true got false
FAIL Property background-image value 'image-set(linear-gradient(black, white) 1x)' assert_true: 'image-set(linear-gradient(black, white) 1x)' is a supported value for background-image. expected true got false
FAIL Property background-image value '-webkit-image-set(linear-gradient(black, white) 1x)' assert_true: '-webkit-image-set(linear-gradient(black, white) 1x)' is a supported value for background-image. expected true got false
FAIL Property background-image value 'image-set(url(http://web-platform.test/example.png) 1x type('image/png'))' assert_true: 'image-set(url(http://web-platform.test/example.png) 1x type('image/png'))' is a supported value for background-image. expected true got false
FAIL Property background-image value '-webkit-image-set(url(http://web-platform.test/example.png) 1x type('image/png'))' assert_true: '-webkit-image-set(url(http://web-platform.test/example.png) 1x type('image/png'))' is a supported value for background-image. expected true got false
FAIL Property background-image value 'image-set(url(http://web-platform.test/example.png) type('image/png'))' assert_true: 'image-set(url(http://web-platform.test/example.png) type('image/png'))' is a supported value for background-image. expected true got false
FAIL Property background-image value '-webkit-image-set(url(http://web-platform.test/example.png) type('image/png'))' assert_true: '-webkit-image-set(url(http://web-platform.test/example.png) type('image/png'))' is a supported value for background-image. expected true got false
FAIL Property background-image value 'image-set(url(http://web-platform.test/example.png) type('image/png') 1x)' assert_true: 'image-set(url(http://web-platform.test/example.png) type('image/png') 1x)' is a supported value for background-image. expected true got false
FAIL Property background-image value '-webkit-image-set(url(http://web-platform.test/example.png) type('image/png') 1x)' assert_true: '-webkit-image-set(url(http://web-platform.test/example.png) type('image/png') 1x)' is a supported value for background-image. expected true got false
FAIL Property content value 'image-set(url('http://web-platform.test/example.png') 192dpi, linear-gradient(black, white) 1x)' assert_true: 'image-set(url('http://web-platform.test/example.png') 192dpi, linear-gradient(black, white) 1x)' is a supported value for content. expected true got false
FAIL Property content value '-webkit-image-set(url('http://web-platform.test/example.png') 192dpi, linear-gradient(black, white) 1x)' assert_true: '-webkit-image-set(url('http://web-platform.test/example.png') 192dpi, linear-gradient(black, white) 1x)' is a supported value for content. expected true got false
Harness: the test ran to completion.
